Add store.bothLayerStoreKinds{,Locked}
This allows to get all the layer stores with a single s.graphLock access.

We do it even in store.CreateContainer, where the read-only layers
are only necessarily conditionally, because getting the read-only
layers is almost always a simple field access, so not having to lock
graphLock again is almost certainly the improvement.

Stand-alone store.getROLayerStores is now unused and has been removed.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>