Reject files mixing v1 and v2 registries.conf, even with empty fields
We already rely on the TOML parser's behavior of setting specified
arays to [], but missing arrays to nil, so just use it to be more strict.

This can help situations like

> [registries.search]
> registries = []
>
> [registries.insecure]
> registries = []
>
> [registries.block]
> registries = []
>
> unqualified-search-registries = ["docker.io"]
>
> [[registry]]
> ...

although only indirectly: From the TOML parser's point of view,
the unqualified-search-registries field is block.unqualified-search-registries
and therefore invisible to us. Conceptually, the same thing could happen with
> [arbitrary-unknown]
> unqualified-search-registries
>
> [[registry]]
> ...

and we wouldn't warn about that.

Even so, the config file author specifying even empty fields is suspect
and suggests a problem to be resolved.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>