Introduce signature/simplesigning.NewSigner
This is a signer.Signer wrapper around the existing simple signing
API, for consistency.

It's just a wrapper which does not substantially change the GPG implementation
approach; doing that might add more benefits (e.g. an immediate failure on an
invalid key ID). That could possibly be improved in the future,
for now the goal is only to migrate copy.Image to using a singer.Signer.

To run tests, we need to introduce a _new_ GPG home directory, because
(go test) runs tests in subpackages in parallel, and the tests in
signature and signature/simple can interfere: a shared gpg-agent will
be unlocked by one test and expected to be locked by another, causing
spurious failures.

So, this adds signature/simplesigning/testdata, but all the GPG files
are symlinks to the original location.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>