Introduce signature/sigstore.NewSigner
This replaces NewSignerWithPrivateKeyFileUnstable, and can be extended
to add more options in the future without having for all of them to be
in the same subpackage (and all dependencies to be included).

The WithPrivateKeyFile option could possibly be in its own subpackage,
but because c/image/copy depends on it, that would not be likely
to remove many dependencies.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>