Remove a direct use of golang.org/x/net/http2
A version of golang.org/x/net/http2 is bundled into
the standard library:
https://github.com/golang/go/blob/master/src/net/http/h2_bundle.go

and here we certainly don't care about the specific version or details.

So just enable use of the built-in one.

Note that this is very unlikely to actually remove the dependency
from binaries, because it is required by google.golang.org/grpc/internal/transport,
and that one has quite a few uses; so most of the benefit right now is just
one less package to get dependency update PRs from.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>