pkg/auth: add `NoWriteBack` option
Add an option to not write back the credentials to the authfile or any
credential helper.  The `/auth` compat endpoint of Podman is currently
not using this code here which ultimately led to normalization errors
surfacing in containers/podman/issues/17571.

The new option allows the endpoint to use this function without writing
back the credentials.

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>