pkg/netns: NewNS() check if file name already exists
The code picks a random name but never checked if this name was already
in use. If the RNG generates the same name again we will end up with
containers sharing the same netns which can be very bad. It is very
unlikely to happen but this the only explanation I find for
https://github.com/containers/podman/issues/17903.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>