pkg/netns: remove old pre go 1.10 workaround
Because we do not not unlock the OS thread it will be killed by the
runtime and not be reused for go 1.10+. Therefore there is no point in
joining the original netns back which never worked as rootless.

Go 1.10 is EOL for years and our code doesn't compile on such old
versions anyway so it is safe to remove.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>