libnetwork/netavark: combine ipvlan macvlan code
Remove the code duplication and share some code between the ipvlan and
macvlan config creation.

Also fixes a problem with the macvlan dhcp selection, we have to set the
driver to dhcp if it is unset and no subnets are given.
Ipvlan does not support DHCP because the interfaces share the same mac
address.

Added tests for both the mcvlan dhcp selection and the ipvlan creation.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>