Use a pointer receiver for internal/set.Set
... to make sure that, and clear to callers that, Add() and Delete()
modify the original object and not a copy.

Luckily(?), this has worked fine, because Set contains a map[],
which is a reference type, so a copy of Set just creates another
reference to the same map.

Either way, fix the method receivers, and add tests to ensure
the code works as expected.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>