.golangci.yml: fix
This fixes the following issue with golangci-lint:

> ERRO Running error: no such linter "cyclop"

Instead of enabling all the linters and then disabling many of
them, start with the default set of enabled linters, and disable
some that are currently giving errors.

This should make upgrading golangci-lint easier, as it maintains a
sensitive set of linters enabled by default (and enabling all linters by
default means we will enable some new and unknown linters every time
we bump golangci-lint version).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>