libnetwork/netavark: fix ipam allocation for custom lease range
LeaseRange allows to set a custom start and end ip for our allocator.
However in this case the allocator ignored that fact the the first ip
(network ip) and the last ip (broadcast address only valid for ipv4)
cannot be used. These ips are special and should never be assigned.

Interestingly the kernel allows to assign the broadcast address to a
interface so it would appear to be working and no error was reported.
Only the attempts to make a connection failed.

Fixes containers/podman#18906

Signed-off-by: Paul Holzinger <pholzing@redhat.com>