libimage: HasDifferentDigest: add InsecureSkipTLSVerify option
To fix a BZ on auto-updates, we need to improve the system tests.
To improve the system tests, we need to use the existing test
infrastructure.  To use the existing test infrastructure, we need
to be able to skip TLS verification in tests.  To be able to skip TLS
verification in the tests, we need a new --tls-verify flag for
podman-auto-update.  And for that to work, we need a way to pass that
down to libimage when checking if there's a new image on the registry.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2218315

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>