Add UnparsedInstanceWithReference
This is useful for combining image data with other reference values,
e.g. to check signatures on a locally-pulled image based on
a remote-registry policy.

This was historically possible to do by simply providing an external
UnparsedInstance implementation; now that we have sigstore signatures
which can't be represent that way, and a private.UnparsedImage,
external users are not allowed to directly access/implement
private.UnparsedImage, so this needs to be an explicitly-provided operation.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>