storage.storageImageDestination.Commit(): leverage image options
When committing an image, pass big data items and the image metadata
string in as part of the set of options we pass to the storage library
when creating the image.

Fall back to setting them individually, as we would have originally,
only if we end up updating an image that was already in local storage.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>