Rename SKOPEO_CI_TAG to SKOPEO_CI_BRANCH
Originally (in https://github.com/containers/image/pull/1272 )
SKOPEO_CI_TAG was also a tag of an on-registry image. That's no longer
the case.

On c/image stable branches, we typically want to test against a corresponding
skopeo stable branch, not against a specific frozen version on that stable branch,
so the variable name has become a bit confusing.

So, rename it to SKOPEO_CI_BRANCH.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>