pkg/config: add (*Config).LoadedModules()
To return absolute paths to modules a config was loaded with.
Knowing the modules is required for conmon's callback to
Podman's cleanup.  Returning them as absolute paths makes
loading the modules a bit faster as it avoids the lookup.

Also drop the attempted performance tune in `Default()` to
accommodate for go's memory model.

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>