overlay: Diffsize: avoid naive diff
Computing the diff size for the rootful overlay storage driver used the
naive diff.  The reasoning was that we have made use of rootful copy-up.
The downside is a mind-blowing performance penalty in comparison to the
rootless case.

Hence, drop the naive diff and only compute the size of the directory
which is absolutely sufficient for the motivating use case of
podman-system-df.  This drops the execution of system-df from 19.3s to
0.11s listing 5 containers and 1 image.

Fixes: github.com/containers/podman/issues/19467
Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>
Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>