golangci.yml: enable all, disable failing linters
golangci-lint has a number of helpful linters and new ones are being
continuously added.  Instead of enabling only a minimal subset of these
linters, enable all by default and - for now - disable failing ones.
This allows us to make use of new linters on updates and to selectively
enable those we deem useful.

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>