ociarchive: Add new `ArchiveFileNotFoundError`
This is for https://github.com/containers/skopeo/pull/2114
which is in turn a dependency of https://github.com/coreos/rpm-ostree/pull/4598

Basically I want to map ENOENT to a clear error, because the build
tooling wants to treat "target image is not present" differently
from "DNS lookup failed" or "we got EPERM".

There's a bit of code motion here because we need to move
the `os.Open()` call before creating a temporary directory.

Signed-off-by: Colin Walters <walters@verbum.org>