Add support for pushing image with unknown digest
Add support to push image with unknown digest and no tag to
a registry. This will be used by farm build to construct a list
of multi arch builds after pushing images built on each node.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>

MH: Backported to release-5.29 branch for Podman 4.9, as this is
required for the `podman farm build` feature. Backporting
features is unusual, but Podman 4.9 is a very unusual release
(based on 4.8 instead of main), which makes this necessary.

I can, at least, say that we do not intend on making a habit out
of this.

Signed-off-by: Matt Heon <mheon@redhat.com>