graphtest: skip tests on insufficient permissions
When a graphdriver cannot be tested because of e.g. `ErrIncompatibleFS`
it is skipped. However this seems not to be the case for permission
error and when running the tests as user they fail with:
```
chown: ... operation not permited
```
This commit changes the test code to account for this as well and
skip when there are insufficient permissions.

Signed-off-by: Michael Vogt <michael.vogt@gmail.com>