pkg/manifests.list.preferOCI(): pay attention to annotations in indexes
Account for the presence of annotations in the OCI version of an index
or list when deciding if the structure should be saved in OCI format.
This requires removing an annotation from the test data so that the
presence of "features" information for one of its instances will be the
deciding factor.

If we don't take those into account, setting annotations in the index
doesn't force the index to be saved in OCI format, so even though none
of the indicators that would force use of the Docker format are present,
we effectively throw away those annotations by using Docker format
anyway.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>