libnetwork/pasta: always pass --quiet
We want to log pasta warnings/errors even if pasta did not fail to
start, however by default it prints a bit to noisy for us. With --quiet
we get the things users should care about.

Unfortunately it still prints to much right now, I will work on some
pasta changes so we can bump it up later.

Note right now --quiet defaults to errors only but it should also do
warnings based on the discussion with Stefano[1]. However this change is
still safe to do right now.

[1] https://archives.passt.top/passt-dev/20240216114304.7234a83f@elisabeth/T/#m42652824644973674e84baf9e0bf1d0e88104450

Signed-off-by: Paul Holzinger <pholzing@redhat.com>