libnetwork/pasta: pass --dns none
We pass --no-map-gw to pasta and when run on a system with only
localhost resolvers pasta cannot find a valid nameserver in resolv.conf
it will print a warning. We do not use the dns functionality so the
warning is just bogus to podman users. Let's disable dns which is only
used for the dhcp option or --dns-forward which we do not use by
default. If a users sets such an option we use them instead.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>